Organize tracer and conntrack header files
