+++ pkg/network/ebpf/c/conn_tuple.h
@@ -1,0 +1,39 @@
+#ifndef __CONN_TUPLE_H
+#define __CONN_TUPLE_H
+
+#include "ktypes.h"
+
+// From include/net/tcp.h
+// tcp_flag_byte(th) (((u_int8_t *)th)[13])
+#define TCP_FLAGS_OFFSET 13
+
+// Metadata bit masks
+// 0 << x is only for readability
+typedef enum
+{
+    // Connection type
+    CONN_TYPE_UDP = 0,
+    CONN_TYPE_TCP = 1,
+
+    // Connection family
+    CONN_V4 = 0 << 1,
+    CONN_V6 = 1 << 1,
+} metadata_mask_t;
+
+typedef struct {
+    /* Using the type unsigned __int128 generates an error in the ebpf verifier */
+    __u64 saddr_h;
+    __u64 saddr_l;
+    __u64 daddr_h;
+    __u64 daddr_l;
+    __u16 sport;
+    __u16 dport;
+    __u32 netns;
+    __u32 pid;
+    // Metadata description:
+    // First bit indicates if the connection is TCP (1) or UDP (0)
+    // Second bit indicates if the connection is V6 (1) or V4 (0)
+    __u32 metadata; // This is that big because it seems that we atleast need a 32-bit aligned struct
+} conn_tuple_t;
+
+#endif // __CONN_TUPLE_H
