+++ pkg/network/ebpf/c/pid_fd.h
@@ -1,0 +1,12 @@
+#ifndef __PID_FD_H
+#define __PID_FD_H
+
+#include "ktypes.h"
+
+typedef struct {
+    __u32 pid;
+    __u32 fd;
+} pid_fd_t;
+
+
+#endif // __PID_FD_H
