+++ pkg/network/ebpf/c/protocols/classification/protocol-classification.h
@@ -4,7 +4,6 @@
 #include "bpf_builtins.h"
 #include "bpf_telemetry.h"
 #include "ip.h"
-#include "tracer-stats.h"
 
 #include "protocols/amqp/helpers.h"
 #include "protocols/classification/common.h"
