+++ pkg/network/ebpf/c/protocols/tls/java-tls-erpc.h
@@ -2,7 +2,6 @@
 #define _JAVA_TLS_ERPC_H
 
 #include "bpf_helpers.h"
-#include "tracer.h"
 #include "tags-types.h"
 #include "https.h"
 #include "port_range.h"
