+++ pkg/network/ebpf/c/skb.h
@@ -16,7 +16,7 @@
 #include <uapi/linux/tcp.h>
 #endif
 
-#include "tracer.h"
+#include "ip.h"
 #include "sock.h"
 #include "ipv6.h"
