+++ pkg/network/ebpf/c/sock.h
@@ -5,7 +5,7 @@
 
 #include "bpf_core_read.h"
 
-#include "tracer.h"
+#include "ip.h"
 #include "ipv6.h"
 #include "netns.h"
