+++ pkg/network/ebpf/c/tracer/events.h
@@ -5,10 +5,10 @@
 #include "bpf_telemetry.h"
 #include "bpf_builtins.h"
 
-#include "tracer.h"
-#include "tracer-maps.h"
-#include "tracer-stats.h"
-#include "tracer-telemetry.h"
+#include "tracer/tracer.h"
+#include "tracer/maps.h"
+#include "tracer/stats.h"
+#include "tracer/telemetry.h"
 #include "cookie.h"
 #include "protocols/classification/tracer-maps.h"
 #include "ip.h"
