+++ pkg/network/ebpf/c/tracer/maps.h
@@ -2,8 +2,9 @@
 #define __TRACER_MAPS_H
 
 #include "map-defs.h"
-#include "tracer.h"
 #include "bpf_helpers.h"
+
+#include "conn_tuple.h"
 
 /* This is a key/value store with the keys being a conn_tuple_t for send & recv calls
  * and the values being conn_stats_ts_t *.
