+++ pkg/network/ebpf/c/tracer/port.h
@@ -1,7 +1,7 @@
-#ifndef __PORT_H
-#define __PORT_H
+#ifndef __TRACER_PORT_H
+#define __TRACER_PORT_H
 
-#include "tracer.h"
+#include "tracer/tracer.h"
 
 #define add_port_bind(pb, pb_map)                                   \
     do {                                                            \
