+++ pkg/network/protocols/kafka/kafka_types.go
@@ -9,7 +9,7 @@
 package kafka
 
 /*
-#include "../../ebpf/c/tracer.h"
+#include "../../ebpf/c/conn_tuple.h"
 #include "../../ebpf/c/protocols/kafka/types.h"
 */
 import "C"
